% Auto-generated: do not edit by hand
\name{bananaNode}

\alias{bananaNode}

\title{BananaNode component}

\description{

}

\usage{
bananaNode(data=NULL, selected=NULL)
}

\arguments{
\item{data}{Lists containing elements 'label'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional). }

\item{selected}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

