% Auto-generated: do not edit by hand
\name{nodeTooltip}

\alias{nodeTooltip}

\title{NodeTooltip component}

\description{
NodeTooltip - A wrapper that displays a tooltip when hovered Built on top of React Flow's NodeToolbar component
}

\usage{
nodeTooltip(children=NULL, className=NULL, offset=NULL, position=NULL,
showOnHover=NULL, style=NULL, tooltipClassName=NULL,
tooltipContent=NULL, tooltipStyle=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The node content}

\item{className}{Character. CSS class for the container}

\item{offset}{Numeric. Offset from the node in pixels}

\item{position}{A value equal to: 'top', 'bottom', 'left', 'right'. Position of the tooltip relative to the node}

\item{showOnHover}{Logical. Show tooltip only on hover (default: true)}

\item{style}{Named list. Inline styles for the container}

\item{tooltipClassName}{Character. CSS class for the tooltip}

\item{tooltipContent}{A list of or a singular dash component, string or number. Content to display in the tooltip}

\item{tooltipStyle}{Named list. Inline styles for the tooltip}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

