% Auto-generated: do not edit by hand
\name{stepEdge}

\alias{stepEdge}

\title{StepEdge component}

\description{
StepEdge - Glass morphism styled stepped edge with sharp corners
}

\usage{
stepEdge(id=NULL, data=NULL, label=NULL, labelStyle=NULL,
markerEnd=NULL, markerStart=NULL, selected=NULL,
sourcePosition=NULL, sourceX=NULL, sourceY=NULL, style=NULL,
targetPosition=NULL, targetX=NULL, targetY=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{Named list. }

\item{label}{Logical | numeric | character | named list | unnamed list. }

\item{labelStyle}{Named list. }

\item{markerEnd}{Logical | numeric | character | named list | unnamed list. }

\item{markerStart}{Logical | numeric | character | named list | unnamed list. }

\item{selected}{Logical. }

\item{sourcePosition}{Character. }

\item{sourceX}{Numeric. }

\item{sourceY}{Numeric. }

\item{style}{Named list. }

\item{targetPosition}{Character. }

\item{targetX}{Numeric. }

\item{targetY}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

