% Auto-generated: do not edit by hand
\name{straightEdge}

\alias{straightEdge}

\title{StraightEdge component}

\description{
StraightEdge - Glass morphism styled straight line edge
}

\usage{
straightEdge(id=NULL, data=NULL, label=NULL, labelStyle=NULL,
markerEnd=NULL, markerStart=NULL, selected=NULL,
sourceX=NULL, sourceY=NULL, style=NULL, targetX=NULL,
targetY=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{Named list. }

\item{label}{Logical | numeric | character | named list | unnamed list. }

\item{labelStyle}{Named list. }

\item{markerEnd}{Logical | numeric | character | named list | unnamed list. }

\item{markerStart}{Logical | numeric | character | named list | unnamed list. }

\item{selected}{Logical. }

\item{sourceX}{Numeric. }

\item{sourceY}{Numeric. }

\item{style}{Named list. }

\item{targetX}{Numeric. }

\item{targetY}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

