% Auto-generated: do not edit by hand
\name{animatedCircleNode}

\alias{animatedCircleNode}

\title{AnimatedCircleNode component}

\description{

}

\usage{
animatedCircleNode(data=NULL)
}

\arguments{
\item{data}{Lists containing elements 'label'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

