% Auto-generated: do not edit by hand
\name{toolbarNode}

\alias{toolbarNode}

\title{ToolbarNode component}

\description{
ToolbarNode - Glass morphism styled node with a configurable toolbar
}

\usage{
toolbarNode(data=NULL, isConnectable=NULL, selected=NULL)
}

\arguments{
\item{data}{Lists containing elements 'label', 'sublabel', 'toolbar', 'toolbarvisible', 'toolbarposition', 'toolbaralign', 'toolbaroffset', 'toolbarstyle', 'style', 'handlestyle', 'targetposition', 'sourceposition', 'showtargethandle', 'showsourcehandle'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional)
  - sublabel (character; optional)
  - toolbar (logical | numeric | character | named list | unnamed list; optional)
  - toolbarvisible (logical; optional)
  - toolbarposition (a value equal to: 'top', 'bottom', 'left', 'right'; optional)
  - toolbaralign (a value equal to: 'start', 'center', 'end'; optional)
  - toolbaroffset (numeric; optional)
  - toolbarstyle (named list; optional)
  - style (named list; optional)
  - handlestyle (named list; optional)
  - targetposition (character; optional)
  - sourceposition (character; optional)
  - showtargethandle (logical; optional)
  - showsourcehandle (logical; optional). }

\item{isConnectable}{Logical. }

\item{selected}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

